[NuGet] Fix UI hang accessing installed packages for .NET Core project
When getting the installed NuGet packages in a .NET Core project
using the public API (PackageManagementProjectOperations) the UI
thread could get locked. This was triggered by the
XamarinGettingStartedProvider looking for a Xamarin.Forms NuGet
package in a .NET Core project. To prevent the UI lock the
DotNetCoreNuGetProject now does not switch to the UI thread to get
PackageReference items from the project. The project
items are stored in a ImmutableList which is thread safe.