[NuGet] DotNetCoreNuGetProject is now a BuildIntegratedNuGetProject
Using the BuildIntegratedNuGetProject as a base class prevents a
restore failure leaving a .NET Core project with package references
being removed since the restore happens before the project is
modified. It also allows the restore on opening the solution to not
run if package restore is not required for a .NET Core project.
Previously the restore was run every time on opening the solution.

Restoring after an install or uninstall of a NuGet package now uses
the DependencyGraphRestoreUtility instead of running MSBuild
/t:restore for the project.