[NuGet] Reload .NET Core project after restore if needed
If the projectname.csproj.nuget.g.targets/.props files are missing
before a restore then when the restore has completed the project
will be reloaded. These NuGet MSBuild files are generated by the
restore process and add extra MSBuild items and properties into
the project and are not available until the project is reloaded.