[NuGet] Show better error when .NET Core NuGet targets are missing
When building a .NET Core project where the generated NuGet targets
are missing you would see the error 'The OutputPath property is not
set for project'. Now when a Build or a Clean is started for a
.NET Core project and the generated NuGet targets are missing a
more informative build error is reported indicating that a package
restore must be run.

Unfortunately it is not possible to automatically run a package
restore when a build has been initiated since the project requires
a reload so the correct MSBuild information is imported from the
new NuGet MSBuild targets that have been generated. Reloading the
project disposes the project and causes the build to be cancelled.