[NuGet] Re-evaluate the .NET Core project after a restore
Previously the .NET Core project would be reloaded as though the file
had been edited externally when a project was created. Also the
reload would occur after a restore if the generated NuGet files
were missing. Now the Project.ReevaluateProject method is used
instead of setting Project.NeedsReload to true.