[DotNetCore] Handle project file changing after having invalid xml
When editing a .NET Core project in the text editor and saving it
with invalid xml the project would show with an error in the
solution window and would be converted to an UnknownSolutionItem.
Further changes in the text editor would not trigger a reload
until the project was right clicked and Reload was selected.

What happens now is that when a project file is modified a check
is made to see if it is currently an UnknownSolutionItem and if so
the project is reloaded. If a project was deliberately unloaded
by right clicking it and selecting Unload then the project will
not automatically be reloaded if the project file is modified in
the text editor.