[DotNetCore] Fix invalid configuration on loading web projects
.NET Core projects that use the Microsoft.NET.Sdk.Web Sdk attribute
in the project file would not have the correct configurations
loaded. The problem is that the .NET Sdk MSBuild import files
use an MSBuildSdksPath property to locate other files. The MSBuild
import files for projects that use Microsoft.NET.Sdk as the Sdk
attribute value do not use this property. To handle this the
MSBuildSdksPath property is added to the MSBuild project file stored
in memory so the imports are correctly resolved. This allows the
configurations to be loaded for the project.