Stop main window disappearing when losing focus
If the breadcrumb bar menu is opened and the user clicks outside of the application, the main window disappears. This is caused by the typehint of the
dropdown menu being set to Menu, which deep in the gdk internals sets
hidesOnDeactivate to YES on the menu window, which somehow seems to propagate
to the main window as well.

Setting the typehint to DropdownMenu instead of just Menu fixes this