[Core] Fix .NET Core project reevaluation error
When loading a project modified in memory the filename was not set
when the project was loaded. .NET Core projects that use the Sdk
attribute require the MSBuildProjectExtension reserved property to
be defined when they are loaded. This property is not set if the
ProjectCollection.LoadProject method is called using an XmlTextReader
and the project reevaluaton fails with the error:

Microsoft.Build.Exceptions.InvalidProjectFileException: The value ""
of the "Project" attribute in element <Import> is invalid. Parameter
"path" cannot have zero length.  /usr/local/share/dotnet/sdk/1.0.0-preview5-004275/Sdks/Microsoft.NET.Sdk/Sdk/Sdk.targets

To fix this the project's XmlTextReader is read by the
ProjectRootElement first, which does not evaluate the project. The
filename can then be set on the ProjectRootElement before the
Project is then created so the MSBuildProjectExtension is set when
the project is then evaluated.