[DotNetCore] Change project templates to use Sdk attribute
Currently the web project compiles but does not show any files
in the solution window since the project file has no wildcard includes
specified and the wildcard items are not imported correctly from the
imports. Adding the wildcard includes to the project breaks the
build since the C# files are added twice.