[DotNetCore] Fix no C# files shown in solution window for web project
The new SDK style web projects have no wildcard includes in the
project file themselves. The wildcard items that are imported are
not currently evaluated correctly which resulted in no files shown
in the solution window. As a temporary workaround an item group
is added to the project in memory with wildcard items for C# and
resource files. Files are then shown in the solution window.