[DotNetCore] Show other files in solution window for web projects
Since a workaround was added for web projects to show C# files
this workaround has been extended to include wildcards for other
files that should be shown in the solution window. The wildcard
items have been added based on what is in the
Microsoft.NET.Sdk.Web.ProjectSystem.props file apart from an item
update which does not cause the project to fail to load.