[DotNetCore] Exclude dot directories from web projects
Updated the wildcard items that are added to a web project in memory
so dot directories, such as .git, are not shown in the Solution
window unless show all files is configured. Also removed the use of
GlobalExcludes which does not seem to be used in the latest dotnet
cli.