[NuGet] Fix duplicate PackageReference on updating
Updating a NuGet package in a .NET Core project was not removing
the existing PackageReference. The DotNetCoreNuGetProject is now
a BuildIntegratedNuGetProject and NuGet does not provide the
uninstall action for this type. Instead only the install action
is given to the BuildIntegratedNuGetProject. So now the existing
PackageReference is removed if it exists when adding a new
NuGet package instead of relying on the UninstallPackageAsync
being called.