[DotNetCore] Update project templates for .NET SDK 1.0.0-rc4-004527
New projects no longer have a toolsVersion attribute.
Removed PackageReference elements which are now implicit from the
project templates, such as Microsoft.NETCore.App and
NETStandard.Library.
The ASP.NET Core project template now uses the Microsoft.AspNetCore
PackageReference instead of adding multiple packages individually.
The ASP.NET Core project template now defines the TargetFramework
property. All the rest have been removed.

To support the changes to these Sdk style projects the following
changes were made:

The OutputType is read from the evaluated properties for Sdk style
projects. This is required since ASP.NET Core projects no longer
define the OutputType directly in the project file but it is defined
by an import. Without this the IDE would not allow the project to
be run.

When creating a new project the toolsVersion was being set to 4.0
by default. Now the DotNetCoreProjectExtension makes sure the
original value is saved for Sdk style projects. This ensures it is
not added for new projects created from the templates.