[DotNetCore] Prevent IDE crash on build
After adding a NuGet package to a .NET Core project and then building
would trigger an exception in the type system which could terminate
the IDE. A document is not part of the workspace error would occur:

FATAL UNHANDLED EXCEPTION: System.ArgumentException: '<Document1a55a90f-f005-4ddc-a2a5-6cead2921bff>' is not part of the workspace.
  at Microsoft.CodeAnalysis.Workspace.CheckDocumentIsInCurrentSolution (Microsoft.CodeAnalysis.DocumentId documentId) [0x00028] in <947408acdcbc4eba93cdb6a19d92fcd1>:0
  at Microsoft.CodeAnalysis.Workspace.OnDocumentTextChanged (Microsoft.CodeAnalysis.DocumentId documentId, Microsoft.CodeAnalysis.Text.SourceText newText, Microsoft.CodeAnalysis.PreservationMode mode) [0x00015] in <947408acdcbc4eba93cdb6a19d92fcd1>:0
  at MonoDevelop.Ide.TypeSystem.MonoDevelopWorkspace+<UpdateFileContent>c__async8.MoveNext () [0x000c7] in monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.TypeSystem/MonoDevelopWorkspace.cs:1080

The problem was that the C# file in the obj directory was added to
the list of documents in the type system, after adding a NuGet package
the project is re-evaluated and this C# file from the obj directory
is not returned from the Project's OnGetSourceFiles this time, and
the type system's list of cached document ids is not updated. At
build time the C# file in the obj directory is modified which causes
the type system to try and update the text for this file. This then
fails since the file is not in the Rosly workspace even though it is
in the list of cached document ids stored in the MonoDevelopWorkspace.
The exception that is thrown about the document not being part of
the workspace is a fatal exception that terminates the IDE.

As a workaround an attempt is made to filter files from the obj
directory out of the project by modifying the exclude that is added
when looking for files. This exclude was using the
BaseIntermediateOutput which does not filter correctly since this uses
the full path and the file included has a relative path. This has
been changed to use the IntermediateOutputPath. This filters the
generated C# file but would not handle other C# files in other
directories inside the obj directory so another change has been made
to explicitly remove any files from inside the obj directory being
returned from the Project's OnGetSourceFiles.