[Core] Fix project builder remoting issue
Fixed the way RemoteProcessConnection handles shutdown. When disconnecting
it first tries to stop gracefully the remote process by sending a Stop message.
If the server doesn't stop after a timeout, it will kill the process.

RemoteProjectBuilder now disposes the connection (which does a disconnect)
instead of disposing the remote builder and relying on it to stop (which didn't
work).