[Core] Add unit tests for MSBuildProject.SetMonoDevelopProjectExtension
The xmlns attribute was not being removed for a Sdk style project
when setting a MonoDevelop project extension. However this did not
affect the file being saved. To make things consistent the xmlns
attribute is now removed if it is an empty string and the project is
an Sdk style project.