[Core] Return empty string namespace instead of null for Sdk projects
The MSBuildProject.Namespace property now returns an empty string
instead of null for Sdk style projects. This makes it easier to use
elsewhere without having to use null checks and replace it with an
empty string. Some XML types automatically support converting a null
namespace to an empty string whilst some do not. So the
MSBuildProject.Namespace does this conversion so other callers do not
have to do the check themselves when using it with a particular XML
API.