[TextEditor] Removed all line changed/inserted/removed events.
They're not supported by the new editor infrastructure but they can
easily be mapped to text replace events. The only downside is that it
requires additional line lookups in some use cases so makes the event
chain a bit slower. (However unnoticeable it's only used in a very few
unimportant places)