[DotNetCore] Allow .NETStandard projects to be referenced
Allow .NET Core projects that target .NETStandard to be referenced
by compatible projects such as Xamarin.iOS and Xamarin.Android
projects. This is done using NuGet's compatibility rules via a project
extension. If the project being referenced has a target framework
of .NETStandard then the NuGet target framework compatibility rule
is used to check if the reference is allowed.