[DotNetCore] Fix broken code completion after referencing a project
Added a workaround for a race condition affecting saving the project
file and the project builder running the ResolveAssemblyReferences
MSBuild task. This would result in the wrong tools version 4.0 being
passed to the project builder and an attempt was made to import the
lib/mono/xbuild/4.0/Microsoft.Common.props file which does not
exist. The workaround is to ensure the project has a tools version
set to 15.0 if it is not set in the original project file, which is
the case for Sdk style projects.