[DotNetCore] Fix OwnerProject null for project ref after reevaluation
When a new project reference is added and then the project is
reevaluated the project reference is replaced with a new instance
but the OwnerProject is not set. This caused code completion to stop
working in the text editor and types were marked with red error
markers.

A new ProjectReference has a BackingEvalItem which has no properties
so on reevaluating a new ProjectReference is created since no match
is found due to the missing properties. On calling Items.SetItem
there is no items added notification since ProjectReference implements
GetHashCode and Equals and the different instances are considered to
be the same. This resulted in a null OwnerProject for the
ProjectReference and when that was written out to xml a null reference
exception was thrown. To fix this the OwnerProject is set after
reevaluation if it is null.