[build] Work around ValueTuple conflict
When compiling, csc prefers the ValueTuple in mscorlib. However, the
Roslyn assemblies reference System.ValueTuple.dll. This _should_ be
fine because Mono also has a System.ValueTuple.dll facade, however by
default System.ValueTuple.dll is local copied and the local copy is
preferred by the runtime, so the types fail to unify and we get a
MissingMethodException.

Work around this by conditionally removing the reference on Unix until
it's properly blacklisted by Mono.

Thanks to Zoltan forhelping to track this down!