[NuGet] Disable reevaluation after restoring packages
Latest Sdk style MSBuild projects do not have MSBuild .targets as
part of NuGet packages so reevaluating the project after adding or
removing NuGet packages is not needed in this case. Disabling this
should also limit the chance of problems due to reevaluation causing
the MSBuildProject to be written to whilst it is being read.