[NuGet] Fix build error after referencing another .NET Core project
When a .NET Core project references a .NET Core Standard library
project the build would fail with an error "Cannot find project info
for 'Lib.csproj'. This can indicate a missing project reference.".
This was due to the wrong path being set as the PackageSpec FilePath.
The path to the project.assets.json file was being used instead of
the project filename. The PackageSpec is used by NuGet to generate the
project.assets.json file during a restore.