[Ide] Fix context menu positioning for some Gtk.TreeView based widgets
Gtk.TreeView contains a windget window and a so called
BinWindow. Both of them raise mouse events with own
pointer coordinates, which may need to be converted.

Additionally the context menu can be requested by
OnPopupMenu without valid EventButton arguments
resulting in NREs if not handled correctly. In this case
we show the context menu at the position of the current
selected item.