[DotNetCore] Fix incompatible project reference build error not reported
With a non .NET Core project referencing a .NET Core Standard library
project if the netstandard target framework version was changed the
build would not fail or succeed based on the current framework version.
The References folder would show correctly an error about an
incompatible framework but the build would not. To fix this the
project builder associated with non .NET Core projects is reloaded if
the .NET Core project is reloaded. This ensures that MSBuild then
correctly reports a build error or success based on the current target
framework in the projects. This problem does not affect .NET Core
projects referencing other .NET Core projects.