[DotNetCore] Handle missing output assembly when running tests
Handle missing output assembly when discovering and running tests.
Test discovery is not started if the output assembly for the project
does not exist. The Unit Tests window allows the tests to be run
when there are no tests so if this is done when there is no output
assembly an error is reported about the missing file which is
displayed in the Test Results window.