Use run button images as template images
Previously, the images for the run button were loaded as-is, which resulted in a black-on-black button that looked disabled. I enabled the template image effect, which is better-- the button is at a weird size now, but at least it doesn't look broken.
Also fixes a bug in the previous commit where I used absolute URLs to load files-- as in, absolute to my home directory on my computer. I'm so, so sorry about that.