[CSharp] Fix on the fly formatter performance issue
This issue is surfaced on cutting and pasting a 19k line file
(TensorFlowSharp with Operations.cs)

The code wasn't formatted in the old variant, causing a ton of changed
spans in the document. Because of this, we ended up recalculating the
whole indentation of the document in NRefactory millions of times (I
could count 30 million before it crashed, 18gb of allocations).

With this change, pasting only does one block change of the text with the
formatted text. This causes invalidation of more lines, but since there
is no proper way of batching text span replacements, this is the only
possible fix I could find.

The code has an extra allocation caused by getting the whole document
text then substring-ing it, because there is no way (afaik) to get a
subtext of a given text.

Common usage might be a bit slower than usual, but this provides a
scalable behaviour