[DotNetCore] Show a dialog if .NET Core SDK cannot be found
On opening or creating a .NET Core project and .NET Core or an SDK
cannot be found then a dialog is displayed indicating it is missing
with a button that currently goes to the Microsoft .NET Core download
page for Mac. If multiple projects are in the solution then only
show the dialog once for the solution.