[DotNetCore] Show a different message on run if .NET Core is missing
When attempting to run an application a message was being shown that
the .NET Core SDK needed to be installed. To run a .NET Core
application only the .NET Core runtime is needed. The message in the
dialog that is shown when .NET Core is not installed has been changed
to be more specific to the action taking place, in this case running
the application. This dialog is now only shown if building before
running is disabled.