[DotNetCore] Show better solution pad error for missing .NET Core SDK
When the .NET Core SDK is missing and the project is a Sdk style
project the Solution window would show a error icon with the tooltip
showing "Invalid configuration mapping". The underlying problem is
that the .NET Core SDK is not installed so the configurations are not
available. So now the tooltip message shows that the .NET Core SDK is
not installed.