[DotNetCore] Show different message if .NET Core SDK is too old
If the .NET Core SDK is installed but is too old or is missing the
new Sdk MSBuild targets a different error message is displayed when
opening the project, in the error tooltip in the Solution window,
and when building the project. This error message indicates that the
SDK is unsupported and a newer version should be installed.
Previously the error message would just indicate that the .NET Core
SDK was not installed which is misleading. If there is no Sdk
directory then the error message the original error message will
still be displayed.