Revert "[CSharp] Fix on the fly formatter performance issue (#1845)"
This reverts commit c2bd602f206281557eecd2ec9b0e46ce2dc1e831.

The main cause of revert is that:
a) the caret offset is screwed up on keypress
b) sometimes more than just what should be formatted is offset

i.e. when typing:
`if`
before a
```
for (int x = 0; x < 100; x++)
	Console.WriteLine (x);
```

The for's inner statement would be indented by 1. Will research for a
better fix.