[AssemblyBrowser] Fix scrolling to type when doing Go To Definition
Before, the code was racy, as it was all attaching a ContinueWith on
LoadingTask, which caused the following semantics:
* On ViewContent loading, it would attach a task to scroll to the loaded
assembly.
* On actual Go To Definition call, it would attach a task that would
scroll to the real type we're searching for.

The order in which these two happen is inverse to the order of
attaching, we would scroll back to the first item, even though we did
all the node expansion and selection work for the real item.

The fix is to only scroll to the assembly node if no scroll to type
happened in the meantime.