[DotNetCore] Fix F# .NET Core projects not failing to build
Added a workaround for F# .NET Core projects when compiling them.
The FSharp.NET.Core.Sdk.targets assume that the dotnet host executable
is shipped with MSBuild which is not the case when using Mono.
The FSharp.NET.Core.Sdk.targets define the path to dotnet using
the MSBuildExtensionsPath so to fix the build this is overridden
when building F# .NET Core projects so it points to the .NET Core
SDK folder.

This will need to be re-addressed if the FSharp.NET.Core.Sdk.targets
are changed. It would be better if they used the MSBuildSDKsPath
which is used by the other .NET Core SDK MSBuild targets.