[DotNetCore] Refresh NuGet packages after target framework changed
When the target framework is changed in project options the NuGet
packages are now restored. To ensure the evaluated properties are
up to date a project reload is triggered. This is used instead of
re-evaluating the project since this seems to cause the save to fail
intermittently.

Making changes in the project options adds extra property groups
when they should not be added.