[DotNetCore] Allow RootNamespace to be updated from project options
If the project did not have RootNamespace specified in the global
property group then changes made to this value would not be saved.
Now a check is made comparing the MSBuildProjectName with the
RootNamespace property and if these are different then the RootNamespace
is saved in the project file. Also added similar support for the
AssemblyName however this seems to currently be saved on a per
configuration basis and is not added to the global properties group.