[DotNetCore] Fix Description not saved in project file if modified
Changing the Description for the project in the project options would
not save the property in the project if it did not exist on loading
the project. Now a check is made on saving to see if the value has
changed from the default to prevent it from being removed when it has
been modified.