[Core] Ignore case when evaluating methods in MSBuild expressions
Expressions of the form $(TargetFrameworkVersion.SubString(1)) are now
supported, where the method case, SubString, does not match the type's
method. This form of expression is used in the .NET Core MSBuild
targets. This fixes the unevaluated expression being shown in the
project options as the compiler's Define Symbol for .NET Core projects.