[Core] Be a little safer about adding/removing tasks from the queue
We should guarantee the task is in the queue *before* the task
to remove it is run. It just keeps the intent clear should the
code be refactored later.

Rewrite the code to use async/await as it makes it a little easier
to reason about as compared to the continuations.