[Core] Allow imported MSBuild items to be added to the project
By default imported MSBuild items are not added to the project. Now
a project or a project extension can indicate that an imported
MSBuild item should be added to the project. This allows .NET Core
projects to add imported files that are not defined directly in the
project itself but from its imports.