[DotNetCore] Fix package restore not run for new projects
Adding a project to an existing solution would not restore the NuGet
packages for the project. Creating a new solution with a .NET Core
project would restore the NuGet packages. A check is now made when
the project is bound to the solution to see if a restore should
occur. Currently adding an existing .NET Core project to the solution
does not trigger a restore since there is no easy way to detect the
difference between this happening and the project being loaded on
opening a new solution.