[DotNetCore] Do not check for tests unless Microsoft.NET.Test.Sdk used
Avoid checking for tests in every project by ensuring the project has
the Microsoft.NET.Test.Sdk NuGet package referenced. Current limitation
with this is that there is no way to show tests if the NuGet package
Microsoft.NET.Test.Sdk is added after the project without reloading the
project. The UnitTestService only rebuilds the tests if a project
reference has changed which is not the case for .NET Core projects
when adding a NuGet package to it. There is a similar problem when
removing the Microsoft.NET.Test.Sdk NuGet package.