[NuGet] Fix license acceptance dialog not shown for .NET Core projects
If a NuGet package has a dependency that required a license acceptance
then the license acceptance dialog was not being shown. It would only
display if the NuGet package being installed required a license and
it would only show the single license.

The problem is that for build integrated NuGet projects, such as .NET
Core projects, the NuGetProjectAction returned from the NuGet
Package manager is a single BuildIntegratedNuGetProjectAction instead
of the individual actions. So a license check was only applied to
the single action instead of the underlying actions that would be
applied when the NuGet package was installed. Now a check is made
to see if a BuildIntegratedNuGetProjectAction is being run and if so
the actions it contains are checked to see if they require a license
to be accepted.