[NuGet] Fix project.assets.json dependency added in wrong section
When installing a NuGet package into a .NET Core project the NuGet
package was being added into the project/dependencies section of the
project.assets.json file. Running a restore after the install would
fix this and put the NuGet package into the project/frameworks/dependencies
section. Visual Studio runs a second restore after installing a NuGet
package which fixes the project.assets.json. So as a workaround a restore
is after the install completes so the project.assets.json file is
regenerated with the correct information.

https://github.com/NuGet/Home/issues/4712