[Core] Support build action changing with wildcard include
When an MSBuildItem is included with a wildcard and then its
BuildAction is changed the project will add a Remove item and an
Include item. The Remove item will remove it from the original
wildcard group.

Wildcard: <Content Include="**\*.txt" />

Changing a .txt file to be an EmbeddedResource:

  <Content Remove="TextFile.txt" />
  <EmbeddedResource Include="TextFile.txt" />

Ordering of the MSBuildItems is not taken into account yet. Visual
Studio groups Remove items together and adds them before the Include
items, which are in their own item groups.