[Core] Fix properties not removed from modified wildcard items
With an MSBuild item, that was originally part of a wildcard include
but has had its build action changed, when a property was set back
to its default value the property was still being saved in the
project file. So if CopyToOutputDirectory was set back to None the
previous value would still be saved in the project file. The problem
was that the original item's properties were not being removed
when the new values were updated.