[Core] Add Update item after wildcard BuildAction change reverted
Changing a build action for a wildcard adds a Remove item and an
Include item. If that build action has property defined and is then
changed back to the original build action the Remove ite and the
Include item are removed and an Update item is added to handle the
property being changed.