[A11y] Don't update the Main menu when creating it.
Updating the main menu is an performance problem, so doing it at start up is not
a great idea. The reason it had been done was because otherwise the
Accessibility Inspector gets no information until the menu has been opened. This might
not be a problem in the real world as screen readers will open the menus first.