[DotNetCore] Support changing .NET Framework version in project options
If the full .NET Framework was targeted by a .NET Core Sdk style project
then it was not possible to change the framework version in project
options. Now the list of installed .NET Frameworks is shown in the
project options in the Target Framework drop down. The target framework
stored uses the shortname version, for example net46 instead of
.NETFramework.