[MSBuild] Restore original global properties after building
Building a project with global properties now restores the original
global properties from the build engine after the build is run.
Previously if the global properties used when building a single
project had the same name as the build engine global properties then
the property would be removed and unavailable for other builds that
did not specify any project global properties.