[DotNetCore] Fix no code completion in new C# file
Fixed bug #53138 - Adding new C# files to a NetCore console project
gets no intellisense until project is unloaded and reloaded.
https://bugzilla.xamarin.com/show_bug.cgi?id=53138

A new C# file added to a .NET Core project would not be returned
to the type system so no code completion would occur in the file
and the type defined in that file would not be available elsewhere.
Added a workaround for now to ensure the files in the project are
returned to the type system.