[Core] Fix project builder issue
When reloading a project builder using in-memory project content, use the
engine's default tools version instead of the tools version specified in the
project. We want to use the latest toolset, even if the project specifies an
older version. This is what would happen when loading the project from a file.

Fixes bug #53126 - Visual Studio for Mac C# editor sprays red everywhere
on initial Android project