[NuGet] Fix PackageTargetFallback being ignored for .NET Core projects
The project.assets.json file generated by a NuGet restore did not
have any imports defined for PackageTargetFallback defined in the
project. This would prevent NuGet packages from being installed.
For example, with an ASP.NET Core project that targets netcoreapp1.0 by
default it is not possible to install Microsoft.Composition 1.30.0
which only contains an assembly for portable-net45+win8+wp8+wpa81. If
a PackageTargetFallback is added, such as the one shown below, then
it should be possible to install the Microsoft.Composition NuGet
package:
   <PackageTargetFallback>$(PackageTargetFallback);dotnet5.6;portable-net45+win8;portable-net45+win8+wp8+wpa81;</PackageTargetFallback>

However this was not working since the imports for the package target
fallbacks were not added to the project.assets.json file on restore.
Now they are being added.