[Core] Support MSBuild 15 on Windows
MSBuild 15 is now used if installed on Windows.

Added a different app.config for the MSBuild runner
MonoDevelop.Projects.Formats.MSBuild.exe on Windows since the
existing one is Mac specific. Added a minimal set of properties in
this app.config to allow a .NET Core project and .NET Framework
project to compile with MSBuild 15. The app.config used on Windows
for MSBuild 15 defines several more properties. The property
definitions that are defined in the app.config used by MonoDevelop
have different values since the values used by Microsoft/MSBuild's
app.config assume that MSBuild.exe is being used when resolving the
properties and uses directories relative to the process being run.