[NuGet] Fix development dependencies not marked in project.assets.json
NuGet packages which are development dependencies need to be marked
with "suppressParent": "All" in the project.assets.json file if they
have PrivateAssets set to All. This was not being done since the
original metadata from the PackageReference was not copied to the
object used to generate the project.assets.json file.