Add support for SDKs fallback search paths
Search paths for MSBuildSDKsPath are now supported, but they are handled
in a special way, since MSBuild really doesn't support fallback paths for
MSBuildSDKsPath. When loading a new project that has a reference to an SDK
we now look for that SDK in any of the registered locations. If a location
is found, the path is provided to the project builder, which sets the
MSBuildSDKsPath env var before loading the project.