[Core] Improve order of inserted Remove, Include and Update items
A Remove item will now be inserted before an Include or Update item
if the MSBuild item has the same type (e.g. Compile). An Include item
will be inserted before an Update item if it has the same type.
Previously the behaviour was that a Remove item would be inserted
before any MSBuild item that had an include.