[NuGet] Fix restore error for .NET Core project referencing shared project
When resolving the shared project in order to generate the
project.assets.json file the project was cast to a DotNetProject and a
null reference exception occurred in the PackageSpecCreator. When
restoring a .NET Core project that references a shared project the
shared project should be ignored and not added to the project.assets.json
file.