[DotNetCore] Do not add imported items to project if not in evaluated items
All items from MSBuildProjects's EvaluatedItemsIgnoringCondition were
added to the project if they were imported and were supported build
actions. However the .NET Core MSBuild tasks allow certain properties
to be set in the project file to false, such as
EnableDefaultCompileItems, which disable the normal imported globs.
However setting these values had no affect on the project and you
could end up with duplicate files. One file from the import and one
from the project itself. Now to avoid this a check is made to ensure
the imported item is in the MSBuildProject's EvaluatedItems before
added it to the project. This also has the side effect of fixing the
duplicate files for .NET Core F# projects even though the F# MSBuild
SDK imports add duplicate F# files to the evaluated items.