[Core] Fix build error when building ASP.NET Core project
Fixed 'The target "Build" does not exist in the project.' error when
building an ASP.NET Core project. The Microsoft.NET.Web.Sdk MSBuild props and targets use the MSBuildSdksPath property when importing other
MSBuild files. The MSBuildSDKsPath property is now explicitly
set as a global property for the project so the MSBuild props and
targets are now correctly imported.