Always set workspace ActiveConfigurationId to a fully expanded one
Always set workspace ActiveConfigurationId to a fully expanded one to
prevent unnecessary reloading of project files. For some projects the
SelectActiveConfiguration first set the ActiveConfigurationId to
non-expanded version (eg. “Debug”) and later called SelectActiveRuntime
which set it again to an expanded one (eg. “Debug|Mixed Platforms”).
This resulted in double reloading of the project files.