[Ide] Fix IWorkbenchWindow.Closing event handling with async operations
If a workbench window contains an unsaved document, we ask the user
to save it and run the save operation asynchronously. This is controlled
by the WorkbenchWindowEventArgs.Cancel property and the result
contols whether we set specific Gtk/GLib events as hanled or not
(i.e. in order to allow or disallow Gtk to close a window).

The first problem here was the async EventHandler for the Closing event
which was not awaited and the window was closed before the
Save operation had a chance to run to completion. This is solved by
an EventHandler delegate returning a Task, making it possible to
await the handlers subscribed to the event.

The second problem with this approach was that Gtk doesn't know
anything about awaiting Tasks and if one of its event handlers
leaves the synchronization context (like the asynchrounous
file saving), it will let the signal pass without waiting for its
handled state to be updated. This is solved by delaying window
closing/destruction (always set the initial delete/close event
handled state and close the window in a continuation task
synchronized with the UI thread)

(fix bug #53089)