[DotNetCore] Cache the .NET Core Sdk path
Now that a condition is being used to check if the project templates
should be displayed in the New Project dialog the .NET Core SDK
path detection was being many times. Now the .NET Core SDK location
and whether it is installed is checked just the single time whilst
the IDE is running.