[DotNetCore] Fix duplicate F# files in Solution window
Creating a new F# .NET Core project would display duplicate F# files
(.fs) in the Solution window. Reloading the project would fix the
duplicates. When creating a new project the FSharp.NET.Sdk NuGet
package's imported MSBuild properties are not available so the
imported F# files from the default Microsoft.NET.Sdk MSBuild imports
are applied. This results in imported F# files as well as those
defined in the project and causes duplicate files in the Solution
window. To fix this imported F# files are ignored and not added to
the project since they will be defined directly in the project.